+++ apps/api/plane/api/urls/__init__.py
@@ -1,6 +1,6 @@
 from .project import urlpatterns as project_patterns
 from .state import urlpatterns as state_patterns
-from .workitem import urlpatterns as workitem_patterns
+from .work_item import urlpatterns as work_item_patterns
 from .cycle import urlpatterns as cycle_patterns
 from .module import urlpatterns as module_patterns
 from .intake import urlpatterns as intake_patterns
@@ -13,7 +13,7 @@ urlpatterns = [
     *asset_patterns,
     *project_patterns,
     *state_patterns,
-    *workitem_patterns,
+    *work_item_patterns,
     *cycle_patterns,
     *module_patterns,
     *intake_patterns,
